Move notification icons and widget icons to separate modules
