+++ core/build.gradle
@@ -40,7 +40,8 @@ dependencies {
     implementation project(':ui:app-start-intent')
     implementation project(':ui:common')
     implementation project(':ui:i18n')
-    implementation project(':ui:png-icons')
+    implementation project(':ui:notifications')
+    implementation project(':ui:widget')
 
     annotationProcessor "androidx.annotation:annotation:$annotationVersion"
     implementation "androidx.appcompat:appcompat:$appcompatVersion"
