+++ settings.gradle
@@ -42,5 +42,6 @@ include ':ui:common'
 include ':ui:echo'
 include ':ui:glide'
 include ':ui:i18n'
-include ':ui:png-icons'
+include ':ui:notifications'
 include ':ui:statistics'
+include ':ui:widget'
