+++ ui/notifications/build.gradle
@@ -4,7 +4,7 @@ plugins {
 apply from: "../../common.gradle"
 
 android {
-    namespace "de.danoeh.antennapod.ui.pngicons"
+    namespace "de.danoeh.antennapod.ui.notifications"
 
     defaultConfig {
         vectorDrawables.useSupportLibrary false
@@ -13,5 +13,9 @@ android {
 }
 
 dependencies {
+    implementation project(':storage:preferences')
+    implementation project(':ui:i18n')
+
     annotationProcessor "androidx.annotation:annotation:$annotationVersion"
+    implementation "androidx.core:core:$coreVersion"
 }
