+++ ui/notifications/src/main/res/drawable/ic_notification_fast_rewind.xml
