+++ ui/notifications/src/main/res/drawable/ic_notification_next_chapter.xml
