+++ ui/notifications/src/main/res/drawable/ic_notification_play.xml
