+++ ui/notifications/src/main/res/drawable/ic_notification_sleep_off.xml
