--- ui/png-icons/README.md
@@ -1,3 +1,0 @@
-# :ui:png-icons
-
-This module provides icon resources. It is necessary to support old Android versions that do not support VectorDrawables. The module generates png versions of all contained icons automatically.  This is only needed for notification icons - for the icons within the app, appcompat handles drawing the vectors.
