+++ ui/widget/README.md
@@ -1,0 +1,4 @@
+# :ui:widget
+
+This module manages the widget. The icons specified in this module are rasterized to png. This is necessary to support old Android versions that do not support VectorDrawables. This is only needed for external icons - for the icons within the app, appcompat handles drawing the vectors.
+
