+++ ui/widget/src/main/res/drawable/ic_widget_fast_rewind.xml
