+++ ui/widget/src/main/res/drawable/ic_widget_pause.xml
