+++ ui/widget/src/main/res/drawable/ic_widget_playback_speed.xml
