+++ ui/widget/src/main/res/drawable/ic_widget_skip.xml
